pyproject.toml
[tool.poetry]
name = "time-audit"
version = "0.1.0"
description = ""
authors = ["David Fernandes"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pandas = "^2.2.2"
ipython = "^8.25.0"
fastapi = "^0.112.0"
uvicorn = "^0.30.0"
python-multipart = "^0.0.9"
sqlalchemy = "^2.0.39"
alembic = "^1.15.2"
pyjwt = "^2.10.1"
pwdlib = "^0.2.1"
argon2-cffi = "^25.1.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
